Group controller helpers by phase
Split dispatch and light execution helpers out of the main controller
loop so controller.go stays focused on actor observation handling.

Rename the sysfs observation entrypoint to make the actor boundary
explicit.
