+++ internal/controller/controller_test.go
@@ -72,7 +72,7 @@ func TestHandleStateChangeTogglesLightRelay(t *testing.T) {
 	}()
 
 	logs := captureLogs(t, func() {
-		handleStateChange(t.Context(), index, commands, sysfs.StateChange{
+		handleSysfsStateChange(t.Context(), index, commands, sysfs.StateChange{
 			Device:   sysfs.Device{Identifier: "di_3_16", Path: "/sys/di_3_16/di_value"},
 			OldValue: sysfs.Off,
 			NewValue: sysfs.On,
@@ -94,7 +94,7 @@ func TestHandleStateChangeLogsRawStateChangeForUnknownDevice(t *testing.T) {
 	index := registry.Build(&entity.Root{})
 
 	logs := captureLogs(t, func() {
-		handleStateChange(t.Context(), index, nil, sysfs.StateChange{
+		handleSysfsStateChange(t.Context(), index, nil, sysfs.StateChange{
 			Device:   sysfs.Device{Identifier: "ro_3_14", Path: "/sys/ro_3_14/ro_value"},
 			OldValue: sysfs.Off,
 			NewValue: sysfs.On,
@@ -121,7 +121,7 @@ func TestHandleStateChangeDoesNotBlockCommandSendAfterCancellation(t *testing.T)
 
 	go func() {
 		defer close(done)
-		handleStateChange(ctx, index, commands, sysfs.StateChange{
+		handleSysfsStateChange(ctx, index, commands, sysfs.StateChange{
 			Device:   sysfs.Device{Identifier: "di_3_16", Path: "/sys/di_3_16/di_value"},
 			OldValue: sysfs.Off,
 			NewValue: sysfs.On,
